fix:  continuous update of blockNumber in graphMetada based on chainId
This commit addresses a bug in the EthereumProvider component where the blockNumber was incorrectly changing based on chainId when calling the switchAppChain function. Since the fetchGraphMetadata function is asynchronous and relies on the chainId state, the blockNumber could being updated with a value from a different chain before the correct graphMetadata is fetched. The interval polling for fetching the graph metadata has been moved to a separate useEffect hook that runs whenever the chainId changes